Refactoring utilities.
