+++ neuroimg/pipeline/04-contact_localization/electrode_clustering.py
@@ -5,19 +5,18 @@ from pathlib import Path
 import nibabel as nb
 import numpy as np
 
-from .utils.io import (
+sys.path.append("../../../")
+
+from neuroimg.base.utils.io import (
     save_organized_elecdict_asmat,
     save_organized_elecdict_astsv,
     load_elecs_data,
 )
-
-sys.path.append("../../../")
-
-from neuroimg.localize_contacts.electrode_clustering.mask import MaskVolume
 from neuroimg.localize_contacts.electrode_clustering.grouping import (
     Cluster,
     CylindricalGroup,
 )
+from neuroimg.localize_contacts.electrode_clustering.mask import MaskVolume
 from neuroimg.localize_contacts.electrode_clustering.postprocess import PostProcessor
 
 
@@ -54,7 +53,7 @@ def load_data(ct_scan, brainmask_ct=None):
 
 
 def compute_electrode_voxel_clouds(
-    elec_coords_mm, ct_img, bm_ct_img=None, output_bm_fpath=None
+        elec_coords_mm, ct_img, bm_ct_img=None, output_bm_fpath=None
 ):
     """Compute electrode voxel clouds via Grouping algo and apply brainmask."""
 
@@ -114,7 +113,7 @@ def apply_threshold_clustering(brainmasked_ct_data, grouped_contacts_vox, **kwar
 
 
 def apply_postprocessing_on_clusters(
-    bound_clusters, sparse_labeled_contacts, ct_affine, **kwargs
+        bound_clusters, sparse_labeled_contacts, ct_affine, **kwargs
 ):
     gap_tolerance = kwargs.get(
         "gap_tolerance", 13
@@ -207,12 +206,6 @@ if __name__ == "__main__":
         help="The output datafile with all the electrode points clustered.",
     )
     parser.add_argument("clustered_voxels_file", help="the voxels output datafile")
-    parser.add_argument(
-        "orgclustered_points_file",
-        help="The output datafile with all the electrode points clustered.",
-    )
-    parser.add_argument("orgclustered_voxels_file", help="the voxels output datafile")
-
     parser.add_argument("binarized_ct_volume", help="The binarized CT volume.")
     parser.add_argument("fs_patient_dir", help="The freesurfer output diretroy.")
     args = parser.parse_args()
